+++ pyro/poutine/escape_messenger.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from .util import NonlocalExit
+from .runtime import NonlocalExit
 
 from .messenger import Messenger
